db1: the telemetry family, and three shapes for costs, ids and nested rows
Six sources move -- token_audit, cost_fold, interaction_events,
guardrail_events, eval and diagnose. Forty-eight operations, and the family is
complete.

Three shapes the wire could not say:

A returned double. db1_cost_fold_total answers with a number of dollars, and
carrying it back as an integer would round a cost at the boundary. It has no
negative failure value to test, so the call happening is the success -- which is
what the domain always meant, since it reports trouble by answering zero.

An array of ids going in. db1_interaction_event_mark_reflected takes int *ids
and a count; the frame carries one cell per id, rendered into storage that
outlives the loop and released after the call. That release was missing from the
repeated-row shape runtime landed with, so this fixes a leak there too.

A nested row. diagnosis_ranking_t's first member is a diagnosis_item_t, so the
reply expands it to its own members -- "hypothesis.content" is how C spells it,
which is why every emitter carries it unchanged.

Two sources do not move with the family.

ensemble reads a template file from a project root and takes a cJSON tree, so
migrating it would put the module in the business of reading paths the daemon
names -- the same question secrets.c raises. It goes to a reserved family of its
own rather than holding five ready sources. diagnostics inspects a database at a
path, like maintenance, so it is infrastructure.

Three things moved out rather than across, all for the same reason: they are not
storage. ie_event_type_name maps an enum to the text the row has always
stored, so db1_interaction_event_record takes that text and the caller converts
-- the delegate_learning precedent. And db1_diagnose_json_full, _json_list and
_render_status query through operations that already exist and format what comes
back; keeping them in the module would have meant giving it a growable string
buffer and a JSON writer to build a rendering nobody stores.

Two declarations were wrong and the fixture said so. Four diagnose adders hand
back the new item's id, and declared as plain writes they would have answered 0
and thrown it away. db1_interaction_event_record answers with its eviction
pass's result rather than an id, which the reply now names honestly.

Survey after: 96 of 113 operations fit the wire, 7 of 13 sources ready.
telemetry no longer appears in the by-family list.

Verified: 51 generator tests; the descriptor, bundle and export suites
(46 + 7 + 7); the FULL unit-tests target (only unit-test-vault-bootstrap, which
passes under env -i); all 60 lint checks; build-integrity; and an end-to-end
case that folds two costs and reads the total as a double, marks two events by
id, and reads a ranking whose first member is a row.

+++ src/db1_client/conversation.c
@@ -1144,7 +1144,7 @@ int db1_user_memory_any()
    int wire_status = call_stage(AIMEE_DB1_OP_USER_MEMORY_ANY, fields, 0, values, caps, 1, NULL);
    if (wire_status != (int)AIMEE_DB1_STATUS_OK)
       return -1;
-   return (int64_t)strtoll(slot0, NULL, 10);
+   return (int)strtoll(slot0, NULL, 10);
 }
 
 int db1_user_memory_upsert(const char *kind, const char *tier, const char *key, const char *content, double confidence, const char *source_session)
